+++ lib/praxis/request_stages/validate_payload.rb
@@ -20,13 +20,13 @@ module Praxis
             request.load_payload(CONTEXT_FOR[:payload])
           rescue Attributor::AttributorException => e
             message = "Error loading payload. Used Content-Type: '#{request.content_type}'"
-            return Responses::ValidationError.new(exception: e, summary: message)
+            return validation_handler.handle!(exception: e, summary: message)
           end
           Attributor::AttributeResolver.current.register("payload",request.payload)
 
           errors = request.validate_payload(CONTEXT_FOR[:payload])
           if errors.any?
-            return Responses::ValidationError.new(summary: "Errors validating payload data", errors: errors)
+            return validation_handler.handle!(summary: "Errors validating payload data", errors: errors)
           end
         end
       end